Organize test/ directory
